refactor(app): clean up the lpc reducer transforms
