+++ app/src/redux/protocol-runs/reducer/transforms/lpc/getNextStepIdx.ts
@@ -1,0 +1,28 @@
+import {
+  LPC_STEPS,
+  ProceedStepAction,
+  StepInfo,
+} from '/app/redux/protocol-runs'
+
+export function getNextStepIdx(
+  action: ProceedStepAction,
+  stepState: StepInfo
+): number {
+  const { toStep } = action.payload
+  const { currentStepIndex, totalStepCount } = stepState
+
+  if (toStep == null) {
+    return currentStepIndex + 1 < totalStepCount
+      ? currentStepIndex + 1
+      : currentStepIndex
+  } else {
+    const newIdx = LPC_STEPS.findIndex(step => step === toStep)
+
+    if (newIdx === -1) {
+      console.error(`Unexpected routing to step: ${toStep}`)
+      return 0
+    } else {
+      return newIdx
+    }
+  }
+}
