+++ app/src/redux/protocol-runs/reducer/transforms/lpc/index.ts
@@ -3,3 +3,4 @@ export * from './handleLwSubstep'
 export * from './handleApplyWorkingOffsets'
 export * from './clearAllWorkingOffsets'
 export * from './updateLPCLabwareInfoFrom'
+export * from './getNextStepIdx'
